+++ src/main/java/com/emosewapixel/pixellib/machines/gui/layout/components/ImageComponent.kt
@@ -15,8 +15,8 @@ open class ImageComponent : IGUIComponent {
     override var height = 166
 
     @OnlyIn(Dist.CLIENT)
-    override fun drawInBackground(mouseX: Int, mouseY: Int, xOffset: Int, yOffset: Int) = texture.render(xOffset + x, yOffset + y, width, height)
+    override fun drawInBackground(mouseX: Double, mouseY: Double, xOffset: Int, yOffset: Int) = texture.render(xOffset + x, yOffset + y, width, height)
 
     @OnlyIn(Dist.CLIENT)
-    override fun isSelected(mouseX: Int, mouseY: Int) = false
+    override fun isSelected(mouseX: Double, mouseY: Double) = false
 }
