+++ src/main/java/com/emosewapixel/pixellib/machines/gui/layout/components/LabelComponent.kt
@@ -19,7 +19,7 @@ open class LabelComponent(val text: String, override val x: Int, override val y:
         )
 
     @OnlyIn(Dist.CLIENT)
-    override fun drawInBackground(mouseX: Int, mouseY: Int, xOffset: Int, yOffset: Int) {
+    override fun drawInBackground(mouseX: Double, mouseY: Double, xOffset: Int, yOffset: Int) {
         val font = Minecraft.getInstance().fontRenderer
         when (location) {
             LabelLocation.CENTER -> font.drawString(text, xOffset - font.getStringWidth(text) / 2f, y.toFloat(), color)
@@ -28,7 +28,7 @@ open class LabelComponent(val text: String, override val x: Int, override val y:
     }
 
     @OnlyIn(Dist.CLIENT)
-    override fun isSelected(mouseX: Int, mouseY: Int) =
+    override fun isSelected(mouseX: Double, mouseY: Double) =
             x < mouseX && mouseX < x + Minecraft.getInstance().fontRenderer.getStringWidth(text) && y < mouseY && mouseY < x + 10
 
     enum class LabelLocation {
